+++ src/main/java/br/com/gerenciamento/dtos/EstablishmentDTO.java
@@ -1,4 +1,4 @@
 package br.com.gerenciamento.dtos;
 
-public record EstablishmentDTO(String name, String document, String address, String telephone, Long amountVacanciesBike, Long amountVacanciesCar) {
+public record EstablishmentDTO(String name, String document, String address, String telephone, Integer amountVacanciesBike, Integer amountVacanciesCar) {
 }
